Restructure Emails context
