+++ lib/code_corps/emails/transmissions/project_approved.ex
@@ -1,32 +1,23 @@
-defmodule CodeCorps.SparkPost.Emails.ProjectUserRequest do
+defmodule CodeCorps.Emails.Transmissions.ProjectApproved do
   import Ecto.Query
 
   alias SparkPost.{Content, Transmission}
   alias CodeCorps.{
-    Presenters.ImagePresenter,
-    Project,
-    ProjectUser,
-    Repo,
-    SparkPost.Emails.Recipient,
-    User,
-    WebClient
+    Project, ProjectUser, Repo, Emails.Recipient, User, WebClient
   }
 
-  @spec build(ProjectUser.t) :: %Transmission{}
-  def build(%ProjectUser{project: project, user: user}) do
+  @spec build(Project.t) :: %Transmission{}
+  def build(%Project{} = project) do
     %Transmission{
       content: %Content.TemplateRef{template_id: template_id()},
       options: %Transmission.Options{inline_css: true},
       recipients: project |> get_owners() |> Enum.map(&Recipient.build/1),
       substitution_data: %{
-        contributors_url: project |> preload() |> url(),
         from_name: "Code Corps",
         from_email: "team@codecorps.org",
-        project_logo_url: ImagePresenter.large(project),
         project_title: project.title,
-        subject: "#{user.first_name} wants to join #{project.title}",
-        user_first_name: user.first_name,
-        user_image_url: ImagePresenter.large(user)
+        project_url: project |> preload() |> project_url(),
+        subject: "#{project.title} is approved!"
       }
     }
   end
@@ -34,11 +25,11 @@ defmodule CodeCorps.SparkPost.Emails.ProjectUserRequest do
   @spec preload(Project.t) :: Project.t
   defp preload(%Project{} = project), do: project |> Repo.preload(:organization)
 
-  @spec url(Project.t) :: String.t
-  defp url(project) do
+  @spec project_url(Project.t) :: String.t
+  defp project_url(project) do
     WebClient.url()
-    |> URI.merge(project.organization.slug <> "/" <> project.slug <> "/people")
-    |> URI.to_string
+    |> URI.merge(project.organization.slug <> "/" <> project.slug)
+    |> URI.to_string()
   end
 
   @spec get_owners(Project.t) :: list(User.t)
@@ -56,6 +47,6 @@ defmodule CodeCorps.SparkPost.Emails.ProjectUserRequest do
   """
   @spec template_id :: String.t
   def template_id do
-    Application.get_env(:code_corps, :sparkpost_project_user_request_template)
+    Application.get_env(:code_corps, :sparkpost_project_approved_template)
   end
 end
