+++ test/lib/code_corps/emails/transmissions/project_approval_request_test.exs
@@ -1,7 +1,7 @@
-defmodule CodeCorps.SparkPost.Emails.ProjectApprovalRequestTest do
+defmodule CodeCorps.Emails.Transmissions.ProjectApprovalRequestTest do
   use CodeCorps.DbAccessCase
 
-  alias CodeCorps.SparkPost.Emails.ProjectApprovalRequest
+  alias CodeCorps.Emails.Transmissions.ProjectApprovalRequest
 
   test "has a template_id assigned" do
     assert ProjectApprovalRequest.template_id
