+++ test/lib/code_corps_web/controllers/organization_invite_controller_test.exs
@@ -38,7 +38,7 @@ defmodule CodeCorpsWeb.OrganizationInviteControllerTest do
       organization_invite_email =
         CodeCorps.OrganizationInvite
         |> Repo.one()
-        |> CodeCorps.SparkPost.Emails.OrganizationInvite.build()
+        |> CodeCorps.Emails.Transmissions.OrganizationInvite.build()
 
       assert_received ^organization_invite_email
     end
